refactor: 重构项目目录结构，统一模块导入路径
1. 移动trigger_word_manager.py到nodes目录下，统一管理节点相关模块
2. 更新所有文件的导入路径，修正相对引用错误
3. 为图片预览接口添加缓存控制头，避免浏览器缓存旧预览图
4. 修复LoRA提示词编码器的文件夹筛选逻辑和图片加载失败处理
5. 新增SDXL工作流示例文件
6. 更新README文档，补充节点说明和使用示例
