+++ nodes/lora_loader_full_node.py
@@ -1,7 +1,7 @@
 import folder_paths
 from comfy_api.latest import io
 
-from .. import trigger_word_manager
+from . import trigger_word_manager
 
 
 class LoRALoaderFull(io.ComfyNode):
