+++ nodes/lora_loader_node.py
@@ -1,7 +1,7 @@
 import folder_paths
 from comfy_api.latest import io
 
-from .. import trigger_word_manager
+from . import trigger_word_manager
 
 
 class LoRALoaderModelOnly(io.ComfyNode):
