+++ nodes/trigger_word_manager.py
